refactor: [task 5] separate rename recovery responsibilities
